requirements/dev.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
alembic==1.10.4
anyio==3.6.2
async-asgi-testclient==1.4.11
asyncpg==0.27.0
autoflake==2.1.1
bcrypt==4.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coverage==7.2.5
databases==0.7.0
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.95.1
flake8==6.0.0
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
mako==1.2.4
markupsafe==2.1.2
mccabe==0.7.0
multidict==6.0.4
orjson==3.8.12
packaging==23.1
pip==23.0.1
platformdirs==3.5.1
pluggy==1.0.0
psycopg2==2.9.6
pyasn1==0.5.0
pycodestyle==2.10.0
pydantic==1.10.7
pyflakes==3.0.1
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-env==0.8.1
python-dotenv==1.0.0
python-jose==3.3.0
pyyaml==6.0
requests==2.30.0
rsa==4.9
sentry-sdk==1.22.2
setuptools==66.0.0
six==1.16.0
sniffio==1.3.0
sqlalchemy==1.4.48
sqlalchemy2-stubs==0.0.2a34
starlette==0.26.1
types-pyasn1==0.4.0.5
types-python-jose==3.3.4.7
typing-extensions==4.5.0
urllib3==1.26.15
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
wheel==0.38.4
